Fixes Instruction SLLI executes even without the 7 MSBs being 0
Change the file immediate.scala to check if the 7 MSBs follow the instruction
spec when the operation is a shift immediate. To test the correct execution,
the file AdeptTester.scala is also edited in order to print the correct values
of PC and registers when a trap is generated. In this file the first `step(1)`
is to set to low the signal load_we and the second `step(1)`, that is
introduced in this commit, is to assert reset. It's important to mention that,
when there is an assembly instruction that will generate a trap, said
instruction should be preceded by two NOPs so that the instructions in the
pipeline may be written to the register file.